Removing ClientOptions and updating PSR-4 folder name
