+++ src/Client.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace CodewarsApiClient;
 
-use CodewarsApiClient\Interfaces\ClientInterface as CodewarsClientInterface;
-use CodewarsApiClient\Interfaces\ClientOptionsInterface;
+use CodewarsApiClient\Interfaces\ClientInterface as CodewarsApiClientInterface;
 use Nyholm\Psr7\Factory\Psr17Factory;
 use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Client\ClientInterface;
@@ -13,19 +12,19 @@ use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Symfony\Component\HttpClient\Psr18Client;
 
-final class Client implements CodewarsClientInterface
+final class Client implements CodewarsApiClientInterface
 {
+    private string $api_key;
+    private string $base_uri = 'https://www.codewars.com/api/v1';
     private ClientInterface $psr18Client;
     private RequestFactoryInterface $psr17Factory;
-    private ClientOptionsInterface $client_options;
-    private string $base_uri = 'https://www.codewars.com/api/v1';
 
     public function __construct(
-        ClientOptionsInterface $options,
+        string $api_key,
         ?ClientInterface $http_client = null,
         ?RequestFactoryInterface $http_factory = null
     ) {
-        $this->client_options = $options;
+        $this->api_key = $api_key;
         $this->psr18Client = $http_client ?? new Psr18Client();
         $this->psr17Factory = $http_factory ?? new Psr17Factory();
     }
@@ -115,13 +114,26 @@ final class Client implements CodewarsClientInterface
         $request_uri = "{$uri}?{$query_string}";
         $request = $this->psr17Factory->createRequest($method, $request_uri);
 
-        foreach ($this->client_options->headers() as $key => $value) {
+        foreach ($this->headers() as $key => $value) {
             $request = $request->withHeader($key, $value);
         }
 
         $response = $this->psr18Client->sendRequest($request);
 
         return $response->getStatusCode() !== 404 ? $this->parse($response) : [];
+    }
+
+    /**
+     * Generate the headers for each request made by the client
+     * 
+     * @return array<string, string>
+     */
+    private function headers(): array
+    {
+        return [
+            'Authorization' => $this->api_key,
+            'Accept' => 'application/json',
+        ];
     }
 
     /**
